+++ src/card/cards/combat_constraints.rs
@@ -148,3 +148,5 @@ pub const STONEWRIGHT: CardDefinitionId = CardDefinitionId(1947);
 pub const DIREGRAF_ESCORT: CardDefinitionId = CardDefinitionId(1948);
 pub const FLOWERING_LUMBERKNOT: CardDefinitionId = CardDefinitionId(1949);
 pub const RELIQUARY_TOWER: CardDefinitionId = CardDefinitionId(1950);
+pub const MUGGING: CardDefinitionId = CardDefinitionId(1951);
+pub const ARREST: CardDefinitionId = CardDefinitionId(1952);